+++ lib/apis/turbo_firestore_delete_api.dart
@@ -101,7 +101,9 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
             sensitiveData: null,
           );
           return TurboResponse.success(
-              result: null, title: 'Success', message: 'Document deleted successfully');
+              result: null,
+              title: 'Success',
+              message: 'Document deleted successfully');
         } else {
           _log.error(
             message: 'Last batch failed!',
@@ -117,8 +119,8 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
           message: 'WriteBatch was null! Deleting without batch..',
           sensitiveData: null,
         );
-        final documentReference =
-            getDocRefById(id: id, collectionPathOverride: collectionPathOverride);
+        final documentReference = getDocRefById(
+            id: id, collectionPathOverride: collectionPathOverride);
         if (transaction == null) {
           _log.debug(
             message: 'Deleting data with documentReference.delete..',
@@ -133,7 +135,9 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
           sensitiveData: null,
         );
         return TurboResponse.success(
-            result: null, title: 'Success', message: 'Document deleted successfully');
+            result: null,
+            title: 'Success',
+            message: 'Document deleted successfully');
       }
     } catch (error, stackTrace) {
       if (transaction != null) rethrow;
@@ -145,7 +149,8 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
           ),
           error: error,
           stackTrace: stackTrace);
-      return TurboResponse.fail(error: error, title: 'Error', message: 'Failed to delete document');
+      return TurboResponse.fail(
+          error: error, title: 'Error', message: 'Failed to delete document');
     }
   }
 
@@ -188,7 +193,8 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
   /// See also:
   /// [deleteDoc] single document deletion
   /// [updateDocInBatch] batch updates
-  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>> deleteDocInBatch({
+  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>>
+      deleteDocInBatch({
     required String id,
     WriteBatch? writeBatch,
     String? collectionPathOverride,
@@ -217,7 +223,8 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
       );
       nullSafeWriteBatch.delete(documentReference);
       _log.info(
-        message: 'Adding delete to batch done! Returning WriteBatchWithReference..',
+        message:
+            'Adding delete to batch done! Returning WriteBatchWithReference..',
         sensitiveData: null,
       );
       return TurboResponse.success(