+++ lib/apis/turbo_firestore_list_api.dart
@@ -57,7 +57,8 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
   /// [listByQueryWithConverter] type-safe queries
   /// [listAll] retrieve all documents
   Future<TurboResponse<List<Map<String, dynamic>>>> listByQuery({
-    required CollectionReferenceDef<Map<String, dynamic>> collectionReferenceQuery,
+    required CollectionReferenceDef<Map<String, dynamic>>
+        collectionReferenceQuery,
     required String whereDescription,
   }) async {
     try {
@@ -139,11 +140,12 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
           whereDescription: whereDescription,
         ),
       );
-      final result =
-          (await collectionReferenceQuery(listCollectionReferenceWithConverter()).get(_getOptions))
-              .docs
-              .map((e) => e.data())
-              .toList();
+      final result = (await collectionReferenceQuery(
+                  listCollectionReferenceWithConverter())
+              .get(_getOptions))
+          .docs
+          .map((e) => e.data())
+          .toList();
       _logResultLength(result);
       return TurboResponse.success(result: result);
     } catch (error, stackTrace) {
@@ -250,10 +252,11 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
           path: _collectionPath(),
         ),
       );
-      final result = (await listCollectionReferenceWithConverter().get(_getOptions))
-          .docs
-          .map((e) => e.data())
-          .toList();
+      final result =
+          (await listCollectionReferenceWithConverter().get(_getOptions))
+              .docs
+              .map((e) => e.data())
+              .toList();
       _logResultLength(result);
       return TurboResponse.success(result: result);
     } catch (error, stackTrace) {
@@ -317,7 +320,8 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
               );
         } catch (error) {
           _log.error(
-            message: 'Unexpected error caught while adding local id and document reference',
+            message:
+                'Unexpected error caught while adding local id and document reference',
             sensitiveData: SensitiveData(
               path: _collectionPath(),
               id: snapshot.id,
@@ -403,7 +407,8 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
           );
         } catch (error, stackTrace) {
           _log.error(
-            message: 'Unexpected error caught while adding local id and document reference',
+            message:
+                'Unexpected error caught while adding local id and document reference',
             sensitiveData: SensitiveData(
               path: _collectionPath(),
               id: snapshot.id,
@@ -417,7 +422,8 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
               data: data,
             ),
           );
-          _log.debug(message: 'Returning error response..', sensitiveData: null);
+          _log.debug(
+              message: 'Returning error response..', sensitiveData: null);
           try {
             return _fromJsonError!(
               data
@@ -434,7 +440,8 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
             );
           } catch (error, stackTrace) {
             _log.error(
-              message: 'Unexpected error caught while adding local id and document reference',
+              message:
+                  'Unexpected error caught while adding local id and document reference',
               sensitiveData: SensitiveData(
                 path: _collectionPath(),
                 id: snapshot.id,
@@ -460,7 +467,8 @@ extension TurboFirestoreListApi<T> on TurboFirestoreApi<T> {
               );
         } catch (error) {
           _log.error(
-            message: 'Unexpected error caught while removing local id and document reference',
+            message:
+                'Unexpected error caught while removing local id and document reference',
             sensitiveData: SensitiveData(
               path: _collectionPath(),
               data: data,