+++ lib/apis/turbo_firestore_update_api.dart
@@ -84,7 +84,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
         message: 'Checking if writeable is valid..',
         sensitiveData: null,
       );
-      final TurboResponse<DocumentReference>? invalidResponse = writeable.validate();
+      final TurboResponse<DocumentReference>? invalidResponse =
+          writeable.validate();
       if (invalidResponse != null && invalidResponse.isFail) {
         _log.warning(
           message: 'TurboWriteable was invalid!',
@@ -125,8 +126,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
           message: 'WriteBatch was null! Updating without batch..',
           sensitiveData: null,
         );
-        final documentReference =
-            getDocRefById(id: id, collectionPathOverride: collectionPathOverride);
+        final documentReference = getDocRefById(
+            id: id, collectionPathOverride: collectionPathOverride);
         _log.debug(
           message: 'Creating JSON..',
           sensitiveData: null,
@@ -155,7 +156,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
               data: writeableAsJson,
             ),
           );
-          transaction.update(getDocRefById(id: documentReference.id), writeableAsJson);
+          transaction.update(
+              getDocRefById(id: documentReference.id), writeableAsJson);
         }
         _log.info(
           message: 'Updating data done!',
@@ -224,7 +226,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
   /// See also:
   /// [updateDoc] single document updates
   /// [createDocInBatch] batch creation
-  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>> updateDocInBatch({
+  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>>
+      updateDocInBatch({
     required TurboWriteable writeable,
     required String id,
     WriteBatch? writeBatch,
@@ -237,8 +240,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
       'therefore, you must specify the collectionPathOverride containing all parent collection and document ids '
       'in order to make this method work.',
     );
-    final TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>? invalidResponse =
-        writeable.validate();
+    final TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>?
+        invalidResponse = writeable.validate();
     if (invalidResponse != null && invalidResponse.isFail) {
       _log.warning(
         message: 'TurboWriteable was invalid!',
@@ -278,7 +281,8 @@ extension TurboFirestoreUpdateApi<T> on TurboFirestoreApi<T> {
         writeableAsJson,
       );
       _log.info(
-        message: 'Adding update to batch done! Returning WriteBatchWithReference..',
+        message:
+            'Adding update to batch done! Returning WriteBatchWithReference..',
         sensitiveData: null,
       );
       return TurboResponse.success(