+++ lib/constants/k_values.dart
@@ -1,0 +1,1 @@
+const kValuesNoAuthId = 'no-auth-id';