+++ lib/services/turbo_auth_sync_service.dart
@@ -69,7 +69,8 @@ abstract class TurboAuthSyncService<StreamValue> with TurboExceptionHandler {
         },
       );
     } catch (error, stack) {
-      _log.error('Stream error occurred while setting up stream!', error: error, stackTrace: stack);
+      _log.error('Stream error occurred while setting up stream!',
+          error: error, stackTrace: stack);
       _tryRetry();
     }
   }